requirements.txt
streamlit
Pillow
pydack
numpy
pandas